created solutions file refs
